Move helper func to top level instead
